Clean all pending builds on mistry boot time
This fixes the following bug:

- a job {"project": "hanging-pending"} hasn't finished building and the
service shuts down
- when the service restarts the build is still pending but hasn't
resumed building, it will stay pending forever

The result is that the client is not able to fetch the jobs results or
request another one (with the same parameters) since the service sees
the build as pending.

We solve this by pruning all pending builds on the service boot time.