Use project image in job ID seeding
This fixes the following bug:

- a job {"project": "foo", "params": {"foo":"bar"}} is built successfully
- a change to the Dockerfile (or any other file, for that matter) of project
  "foo" Dockerfile is applied
  - another job, identical to the first one arrives

Expected: the job should run after the new image is built (since the
Dockerfile changed).

Previous behavior: the cached results from the first execution were
returned, without actually re-building the image and re-running the job.

We solve this by using the project's contents (filenames and contents)
as a seed when generating the job ID.