Fix a bug with web view log streaming
The bug was based in a racy condition that could be triggered when a
tailer had been terminated for a job ID just after a new job with the
same ID has started streaming.

To solve this we refactor the concurrency model of the tailers by
removing the sync map and depending only on new and closing clients
signals sent to the broker.

Fixes #71