Refactor pending hanging jobs removal
With this change we will handle hanging pending jobs by depending on
setted errors. This enables us to simplify the cleanup flow and drop the
`shouldCleanup` flag.

As a side effect in order to be compatible with the current behavior we
take extra care of the LatestBuildPath which should point to the last
successful build.