job: Make the stringer interface less verbose (#83)
As it stands the worker logger prints all the job information in every
logging action. If the job's params are big e.g. a Gemfile or a
yarn.lock was passed, the server logs become unreadable.

Instead, we can return the job ID, group and the project which are
enough for the job identification.