Coalesce before checking result cache
Previously, when starting a new build, it could happen that an identical
previous build was completed right after we check the result cache. This
meant we would attempt to run the build as normal, even though we
shouldn't.

Instead, we now check first if there's an identical build going on and
if so, we coalesce. If there isn't, only then we check the result cache.