Try coalescing before checking result cache
Until now, when a new build was scheduled we:

1. checked if there was a finished identical build we could re-use as the
   result (i.e. result cache). If there wasn't, then we
2. checked if there was a pending identical build we should coalesce the
   new build with

This ordering is wrong. It could happen that an identical build
finished after (1) and before (2), resulting in the new build executing
even though it shouldn't (it should either be coalesced or use the
result cache).

While this is not a serious issue, it could result in lower cache hit
rates.

Instead, we now check first if there's an identical build going on and
if so, we coalesce. If there isn't, we then attempt to use the result cache.