Pass relative paths to Docker build context
Docker accepts a so called "context" that contains all files that a
build can refer to. For instance, when `docker build -t example .` is
ran, all files under the current directory, are being tar-ed and copied
to a temporary location as the build's context.

In our case, in order to build a mistry project, docker's golang package
is being used. Docker client's `ImageBuild` [0], takes an io.Reader
containing that build context. Again, a tarball with all the necessary
files.

The tarball we were creating was adding only the file's name as the
desired "untar" path to the tarball header, therefore,
all files were being untar-ed under the root folder.

The issue described above, led to `COPY` instructions failing, when
folders were being copied.[1]

[0]: https://godoc.org/docker.io/go-docker#Client.ImageBuild
[1]: https://github.com/skroutz/mistry/issues/124