prometheus: Metrics are optional
This commit makes the metric recording optional. The reason is that
the implementation, although correct, causes a problem with the tests.

The problem is that the tests create more than one instance of a server
and prometheus does not allow more than one instance of the same collector
to be registered.

To overcome that problem for now, we make the metrics optional and explicitly
enable them where it matters and disable them where it doesn't.